update migrations and bash script

+++ importdb.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-rm latest.dump
+if [ -f latest.dump ]; then
+  rm latest.dump
+fi
 heroku pg:backups:capture --app film-movement
 heroku pg:backups:download --app film-movement
 pg_restore --verbose --clean --no-acl --no-owner -h localhost -d film-manager_development latest.dump
